fix(@angular-devkit/build-angular): add missing declaration types in build_angular schemas.
Fixes issue #13388
