+++ packages/angular_devkit/build_angular/src/karma/schema.d.ts
@@ -5,7 +5,12 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-import { BrowserBuilderSchema } from '../browser/schema';
+import {
+  AssetPatternObject,
+  BrowserBuilderSchema,
+  CurrentFileReplacement,
+  NormalizedSourceMaps,
+} from '../browser/schema';
 
 export interface KarmaBuilderSchema extends Pick<BrowserBuilderSchema,
   'assets' | 'main' | 'polyfills' | 'tsConfig' | 'scripts' | 'styles' | 'stylePreprocessorOptions'
