+++ packages/angular_devkit/build_angular/src/server/schema.d.ts
@@ -8,7 +8,10 @@
 
 import {
   BrowserBuilderSchema,
+  CurrentFileReplacement,
   FileReplacement,
+  NormalizedOptimization,
+  NormalizedSourceMaps,
   OptimizationObject,
   SourceMapOptions,
 } from '../browser/schema';
